get redux store data in react component
