+++ src/App.css
@@ -6,6 +6,31 @@
   height: 40vmin;
   pointer-events: none;
 }
+/* header style */
+.header{
+  background-color: orange;
+  height: 100px;
+}
+.cart-div{
+  float: right;
+  margin: 30px;
+  position: relative;
+}
+.cart-div img{
+  width: 30px;
+}
+.cart-div span{
+  background-color: green;
+  border-radius: 100%;
+  height: 20px;
+  width: 20px;
+  position:absolute;
+  top:-10px;
+  right:0px;
+  color: #fff;
+}
+
+/* end header style */
 
 @media (prefers-reduced-motion: no-preference) {
   .App-logo {
