+++ src/App.js
@@ -1,19 +1,12 @@
-import logo from './logo.svg';
 import './App.css';
-import { addToCart } from './redux/action';
-import { useDispatch } from 'react-redux'
+import Header from './components/Header';
+import Main from './components/Main';
 
 function App() {
-  const dispatch = useDispatch();
-  const product = {
-    name: 'i Phone',
-    category: 'mobile',
-    price: 10000,
-    color: 'red'
-  }
   return (
     <div className="App">
-      <button onClick={() => dispatch(addToCart(product))}>Add To Cart</button>
+      <Header />
+      <Main />
     </div>
   );
 }
